@@ -6,44 +6,19 @@ use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Exception\MissingOptionsException;
 
 /**
- * Class ZipCode
  * @Annotation
  * @Target({"PROPERTY", "METHOD", "ANNOTATION"})
  * @author Martin Schindler
- * @package ZipCodeValidator\Constraints
  */
 class ZipCode extends Constraint
 {
-    /**
-     * @var string
-     */
-    public $message = 'This value is not a valid ZIP code.';
+    public string $message = 'This value is not a valid ZIP code.';
+    public ?string $iso = null;
+    public ?string $getter = null;
+    public bool $strict = true;
+    public bool $caseSensitiveCheck = true;
 
-    /**
-     * @var string
-     */
-    public $iso;
-
-    /**
-     * @var string
-     */
-    public $getter;
-
-    /**
-     * @var bool
-     */
-    public $strict = true;
-
-    /**
-     * @var bool
-     */
-    public $caseSensitiveCheck = true;
-
-    /**
-     * ZipCode constructor.
-     * @param mixed $options
-     */
-    public function __construct($options = null)
+    public function __construct(mixed $options = null)
     {
         if (null !== $options && !is_array($options)) {
             $options = array(
@@ -54,7 +29,7 @@ class ZipCode extends Constraint
         parent::__construct($options);
 
         if (null === $this->iso && null === $this->getter) {
-            throw new MissingOptionsException(sprintf('Either the option "iso" or "getter" must be given for constraint %s', __CLASS__), array('iso', 'getter'));
+            throw new MissingOptionsException(sprintf('Either the option "iso" or "getter" must be given for constraint %s', __CLASS__), ['iso', 'getter']);
         }
     }
 
