@@ -1,11 +1,14 @@
 <?php
 
+namespace ZipCodeValidator\Tests\Constraints;
+
+use PHPUnit\Framework\TestCase;
 use Symfony\Component\Validator\Exception\MissingOptionsException;
 use ZipCodeValidator\Constraints\ZipCode;
 
-class ZipCodeTest extends \PHPUnit\Framework\TestCase
+class ZipCodeTest extends TestCase
 {
-    public function testMissingOptionsExceptionWhenIsoAndGetterIsEmpty()
+    public function testMissingOptionsExceptionWhenIsoAndGetterIsEmpty(): void
     {
         $this->expectException(MissingOptionsException::class);
         $constraint = new ZipCode(null);
